requirements.txt
moondream
opencv-python
pillow
torch
torchvision
